+++ imagine/include/imagine/logger/logger.h
@@ -2,17 +2,17 @@
 
 #include <imagine/config/build.h>
 #include <imagine/config/imagineTypes.h>
-#include <stdarg.h>
-#include <imagine/util/ansiTypes.h>
 #include <imagine/util/builtins.h>
+#include <stdarg.h>
+#include <stdint.h>
 #include <stdbool.h>
 
 enum { LOGGER_ERROR, LOGGER_WARNING, LOGGER_MESSAGE, LOGGER_DEBUG_MESSAGE };
 
-static const uint loggerMaxVerbosity = LOGGER_DEBUG_MESSAGE;
-extern uint loggerVerbosity;
+static const uint8_t loggerMaxVerbosity = LOGGER_DEBUG_MESSAGE;
+extern uint8_t loggerVerbosity;
 
-typedef uint LoggerSeverity;
+typedef uint8_t LoggerSeverity;
 
 CLINK void logger_init() __attribute__((cold));
 CLINK void logger_setEnabled(bool enable);
@@ -24,10 +24,10 @@ CLINK void logger_vprintf(LoggerSeverity severity, const char* msg, va_list arg)
 #define logger_printfn(severity, msg, ...) logger_printf(severity, msg "\n", ## __VA_ARGS__)
 
 // some shortcuts
-static const uint LOG_M = LOGGER_MESSAGE;
-static const uint LOG_D = LOGGER_DEBUG_MESSAGE;
-static const uint LOG_W = LOGGER_WARNING;
-static const uint LOG_E = LOGGER_ERROR;
+static const uint8_t LOG_M = LOGGER_MESSAGE;
+static const uint8_t LOG_D = LOGGER_DEBUG_MESSAGE;
+static const uint8_t LOG_W = LOGGER_WARNING;
+static const uint8_t LOG_E = LOGGER_ERROR;
 
 #ifndef LOGTAG
 #define LOGTAG
